0.16

+++ Experio/Source/Runtime/Rendering/Managers/FontManager.cpp
@@ -83,6 +83,9 @@ void FontManager::OnFontDeleted(uint16_t slotIndex)
 	slot.refCount--;
 	if (slot.refCount == 0)
 	{
+		if (slot.persistInMemory)
+			return;
+
 		slot.DeleteFont();
 		slotNames[slotIndex] = "";
 		if (nextAvailibleSlot > slotIndex)
@@ -96,12 +99,18 @@ void FontManager::FindNextAvailibleSlot()
 {
 	for (int i = 0; i < MAX_FONTS; i++)
 	{
-		if (slots[i].data == nullptr && slots[i].refCount <= 0)
+		if (slots[i].data == nullptr && slots[i].refCount <= 0
+			&& slotNames[i].size() == 0)
 		{
 			nextAvailibleSlot = i;
 			return;
 		}
 	}
+}
+
+bool FontManager::CanStoreFont()
+{
+	return NumLoadedFonts() < MAX_FONTS - 1;
 }
 
 void FontManager::Init()
@@ -124,7 +133,12 @@ FontRef FontManager::GetFont(const std::string & str)
 	return FontRef(nullptr, MAX_FONTS);
 }
 
-FontRef FontManager::LoadFont(const std::string & str)
+FontRef FontManager::GetFont(uint16_t slotIndex)
+{
+	return FontRef(slotIndex);
+}
+
+FontRef FontManager::LoadFont(const std::string & str, bool persistInMemory)
 {
 	for (uint16_t i = 0; i < MAX_FONTS; i++)
 	{
@@ -134,17 +148,20 @@ FontRef FontManager::LoadFont(const std::string & str)
 		}
 	}
 
-	if (slots[nextAvailibleSlot].data != nullptr)
+	if (slots[nextAvailibleSlot].data != nullptr 
+		|| slotNames[nextAvailibleSlot].size() > 0)
 	{
 		FindNextAvailibleSlot();
 	}
 
-	slotNames[nextAvailibleSlot] = str;
-	slots[nextAvailibleSlot].data = FontReader::ReadFile(str.c_str());
-	slots[nextAvailibleSlot].refCount++;
-	uint16_t lastAvailibleSlot = nextAvailibleSlot;
+	uint16_t slotId = nextAvailibleSlot;
+
+	slotNames[slotId] = str;
+	slots[slotId].persistInMemory = persistInMemory;
+	slots[slotId].data = FontReader::ReadFile(str.c_str());
+	slots[slotId].refCount++;
 	FindNextAvailibleSlot();
-	return FontRef(lastAvailibleSlot);
+	return FontRef(slotId);
 }
 
 bool FontManager::IsFontLoaded(const std::string & str)
@@ -162,6 +179,17 @@ bool FontManager::IsFontLoaded(const std::string & str)
 std::string FontManager::GetNameOfFont(const FontRef & ref)
 {
 	return slotNames[ref.fontID];
+}
+
+size_t FontManager::NumLoadedFonts()
+{
+	size_t numLoadedFonts = 0;
+	for (uint16_t i = 0; i < MAX_FONTS; i++)
+	{
+		if (slots[i].data != nullptr)
+			numLoadedFonts++;
+	}
+	return numLoadedFonts;
 }
 
 size_t FontManager::SizeOfLoadedFonts()
